feat: replace --part option with switch-provided port map
Prior to this change, the port map was hardcoded in a look up table. The
default port map was for the VSC5640EV.  If the user attempted to
configure the EV23X71A LAN969x 24-port EVB with the default `--part`,
the switch would throw an unhelpful error about unrecognized ports due
to the use of the wrong port map.

This change improves this UX by removing the hardcoded port maps in
favor of querying the switch for the port map.  This allows any switch
to be targeted without the user having to know the switch part number
and without having to pass additional options.

This change also modifies how the `--dry-run` option works.  When
`--dry-run` is present, we may not have access to the switch and
therefore may not be able to query the switch for port map information.
And we no longer have that information hard coded.  So instead, we now
use a dummy port map. The following alternative was also considered:
attempt to query switch for port map and fail over to dummy port map
when no response. This alternative was abandoned since it could not
easily be made to fail fast. DNS alone took multiple seconds to fail.
Getting this to fail fast would have required significant additional
complexity.

BREAKING CHANGE: This change removes the `--part` option since it is no
longer needed.  This option was never widely used so this shouldn't
break anyone.